feat: reduce vaults initial load
